feature/en proceso de corregir las colas de rabbitMQ

+++ backEnd/gestion-asignacion/referencias/models.py
@@ -12,7 +12,8 @@ class Programa(models.Model):
 
 
 class Estudiante(models.Model):
-    est_codigo = models.AutoField(primary_key=True)
+    # CAMBIO: Se ajusta a IntegerField para coincidir con el modelo original.
+    est_codigo = models.BigIntegerField(primary_key=True)
     est_nombre = models.CharField(max_length=120)
     est_apellido = models.CharField(max_length=120)
     est_correo = models.EmailField(max_length=254, unique=True)
