+++ backEnd/gestion-formulario/gestion_estudiantes/models.py
@@ -2,7 +2,7 @@ from django.db import models
 from gestion_electivas.models import Programa
 
 class Estudiante(models.Model):
-    est_codigo   = models.AutoField(primary_key=True)
+    est_codigo = models.BigIntegerField(primary_key=True)
     est_nombre   = models.CharField(max_length=120)
     est_apellido = models.CharField(max_length=120)
     est_correo   = models.EmailField(max_length=254, unique=True)
